Refactor init to be asynch
With the introduction fo thr AwsEcsDetector, which does outbound I/O,
we need to change the startup logic of the Lumigo distro to be asynch.

This change introduces a const Promive<void> sdkInit that clients can
chain against to know when the SDK is initialized, or if an error has
occurred.