Provide access to TracerProvider over init promise
Provide means of accessing the underpinning BasicTracerProvider
through the resolution of the initialization promise, which enables
access to the resource (which we need for tests), adding more
span exporters (in case Lumigo must co-exist with other APMs) and,
most importantly, a way to flush span to Lumigo before the process
terminates.

Additionally, implement the rejection of the init promise on failed
initialization as originally documented.